requirements.txt
pytest
Click
Pillow